refactor(item-group): rename Header to Title and move ExternalRecipeEditModal to its own file
